requirements.txt
streamlit
pandas
numpy
plotly
scikit-learn
pyyaml
joblib
pyttsx3
openai
